Fixed 236240
Sparkle now detects if the preferences for automatic update checks or the time interval change mid-cycle.

If your product is a non-.app, you need to clue Sparkle in on the change by calling [[SUUpdater sharedUpdater] updatePreferencesChanged]. Sadly, there's no way to use NSUserDefaultsController to get to preferences which aren't managed by the host app.