Fixed 243533
This should fix the .tar.bz2 and .tar.gz cases. Basically, Launch Services isn't smart enough to see both parts of the extension. When it sees a .bz2 or a .gz, I assume it's a .tar.bz2 or a .tar.gz, because that's what makes sense. Don't use plain .bz2 or .gzs!