Delegate methods are called before and after the update dialog is shown
For now only the modal dialog, that appears when an insecure update is found, provides delegate
methods an application can respond to, e.g. hiding / minimizing windows to make place for the popup
notification.
With this feature addition it is also possible to recognize when the update dialog appears or is
closed after initiating a UI based update check. In this way an application can respond to these events.
This is necessary for the development of a client application and might be of interest for other users too.