Added delegate methods to all UI bits
This is still incomplete and not perfect yet.

In particular, these delegate methods ought to perhaps be grouped in a separate class.

The UI code that is currently done should be treated the same manner as a custom UI

Relaunching Sparkle should give customization over showing the UI of the autoupdater app