Making Sparkle more XPC friendly
We support NSSecureCoding for SUAppcastItem and SUHost

Callbacks from XPC <-> Host app now async to main GCD queue.
This is needed for everything to be thread safe.

Todo: NSDate parsing sucks and is temporarily disabled.
In particular, we're using deprecated methods to read the dates,
which don't support NSSecureCoding. The other problem is that the dates are
really useless technically and we shouldn't sort by them.