ATS cannot be reliably checked against anymore
An app can contain a XPC service with restricted entitlements and own ATS exceptions.
Indeed, Sparkle now is a framework that embeds two XPC services that include download entitlements.
If ATS is disabled, it'd have to be disabled in those services, not within the root of the application.

So we can't keep this detection because it will generate false positives. And it's not really our responsibility anyway.