Add release build configuration that allows arbitrary HTTP loads
In debug mode we always build with arbitrary http loads, but in release mode we
only want to build sparkle-cli and SparkleTestApp with arbitrary http loads.

An ATS error was added back if arbitrary HTTP loads are not allowed and if the feed URL is http://

The "distribution" target/scheme is now not very up to date because it builds everything under Release configuration.