Add installation did finish event. This is not complete.
For sparkle-cli and out-of-process updaters, I added a did finish installing event.

I also have logic dealing with whether an update can be silently installed or not.

I made closing status windows before app termination a reponsibility for the user driver now, not
the installer drivers.

I improved checking for updates before starting the updater, if Sparle didn't ask the user for permission yet.

Currently, Autoupdate can persist in the background even when the app is terminated,
if we don't have sufficient privilleges to install right away for example, or if an update isn't silent.
After much thinking, I need to ditch this strategy because:

1. An app can be manually replaced and so the installer would have to monitor that
2. We leave a zombie process running