Installer progress: change the dock icon to the bundle being updated
This way we don't have to include an icon for the installer progress app. There is a minor issue where the dock icon changes back before termination. However, this app should show up in rare cases in the first place, so I still think this is a good trade off.

For most applications, it makes more sense to use the app's icon than Sparkle's. That and we can save some disk space usage.