Delay check for -isRunningOnReadOnlyVolume until we know if update available or not
This commit postpones the check for -isRunningOnReadOnlyVolume until
*after* Sparkle has already discovered a valid update and the user has
chosen to download the update (either via the UI or via automatic
upgrade driver). I think there are at least two advantages to doing
this check later rather than sooner:

1. For apps that might be affected by Gatekeeper Path Randomization
(#832), this change allows the app to still get delegate callbacks to
know when an update is available (even if it can’t be applied). If it
chooses, the app could then prompt the user to take some action. For
example, moving the app out of ~/Downloads—or offering to do it on the
user’s behalf via something like LetsMove.framework (which now works
again in 10.12 beta 3). Without this commit, the update fails before
the app has a chance to let the user know they’re missing out on an
update.

2. This commit allows -[SUUpdater checkForUpdateInformation] to
function properly even if app on read-only volume.