Correct
The old syntax with single or double dash and no spaces on either side of the dash is very non-standard English (and Afrikaans; I cannot speak for other languages). I have changed it to bring it in line with the same string on WinSparkle, https://winsparkle.org/ (I was unable to locate the source file on Github, but this screenshot shows the difference in formulation.