If no installable item can be found in appcast, return the item with the highest version that can be installed on a different version of the OS
We want to be able to show the user a message that he needs to upgrade his OS if he wants to receive updates in the future.

For this we need to return a suitable app cast item from bestItemFromAppcastItems:getDeltaItem:withHostVersion:comparator: that we can pass to the upper layer, which does already check if the returned item is installable.

So this should not change the current behavior that the user experiences.