Show/Hide skip/later buttons based on whether update is critical / automatic updates enabled.
Removed keys for setting buttons to hidden using plist entries.
Instead, the skip-button is now hidden for critical updates, and the later-button is hidden if automatic updates are disabled.
Note that the user can still choose to just close the window by clicking the red dot at the top of the window.