Add support for selectively choosing what updates are info-only
Example:
<sparkle:informationalUpdate>
    <sparkle:version>1.5.0</sparkle:version>
    <sparkle:version>1.5.1</sparkle:version>
</sparkle:informationalUpdate>

Means that updating only from versions 1.5.0 and 1.5.1 are shown an informational update (provided there is an info link).

<sparkle:informationalUpdate></sparkle:informationalUpdate>

Means that updating from any version gets an informational update (provided there is an info link).

No presence of this new element falls back to current behavior in Sparkle where an informational update is shown if an enclosure is missing but an info link is provided.