Add generate_appcast options for appcast selection features
We can now supply for new update items:
* Link
* Channel name
* Last major version / minimum autoupdate version
* Phased rollout interval
* Last critical update version
* Originating informational update versions

We can also specify the maximum number of delta items to generate, and fixed a index-off-by-1 error.

Lastly we can specify which versions to create new items for. This allows applying flags for only specific versions, or allows adding old versions that do not meet the most recent updates threshold (think for updates for a prior OS or for a prior major version).

We do not override the minimum system version element on already existing items. Apps don't strictly need to specify LSMinimumSystemVersion and this allows users specifying it themselves.

The help page for generate_appcast was expanded with how the archive directory looks like, how release notes work, how old entries are kept in tact, how extracted archives are cached..

We now write out the number of new updates written and the number of existing items updated, to make the generated result be less confusing.