Add and adopt updaterDidFinishUpdateCycle:error: delegate method
The SPUCommandLineUserDriver strictly drives user interface now and does not decide to exit in the middle of an update cycle.

We now exit with different exit statuses if we're not allowed to update eg because major upgrades aren't allowed, new updates aren't found, installer interaction isn't allowed, update permission can't be automatically granted, and so on.